+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 - Expanded the `validation` to also validate __negated__ export-ignore directives. Closes [#70](https://github.com/raphaelstolt/lean-package-validator/issues/70).
+- New `--flavour` option for the `create` command which influences the `.gitattributes` generation. Closes [#71](https://github.com/raphaelstolt/lean-package-validator/issues/71).
 
 ## [v5.9.1] - 2026-05-17